# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "trip API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/trips/{trip_id}/generate": {
      "post": {
        "tags": [
          "generation"
        ],
        "summary": "Start Generation",
        "description": "일정 생성을 시작하고 **즉시** job_id 를 반환한다.\n\nNFR-1 — 최대 60초가 걸리므로 HTTP 응답을 붙잡지 않는다.",
        "operationId": "start_generation_api_trips__trip_id__generate_post",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripSpecIn"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "tags": [
          "generation"
        ],
        "summary": "Get Job",
        "description": "진행 상태 폴링 (WF-2 6단계).\n\n`partial` 은 실패가 아니라 품질 저하다 (DD-23). 프론트는 이를 구체적으로\n알려야 한다 — \"3곳을 찾지 못했습니다\", \"이동시간이 추정치입니다\".",
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/places/search": {
      "get": {
        "tags": [
          "places"
        ],
        "summary": "Search Places",
        "description": "FR-6 — 지역검색은 1회 5건이 상한이라 `page` 로 이어 받는다 (CON-2).",
        "operationId": "search_places_api_places_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "title": "Q"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedPlacesOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/places/content": {
      "get": {
        "tags": [
          "places"
        ],
        "summary": "Place Content",
        "description": "FR-20·21 — 추천 콘텐츠.\n\nBR-40 — 근거 블로그가 3건 미만이면 `highlights` 는 빈 목록이다.",
        "operationId": "place_content_api_places_content_get",
        "parameters": [
          {
            "name": "trip_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 36,
              "title": "Trip Id"
            }
          },
          {
            "name": "item_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 36,
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceContentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/places/suggestions": {
      "get": {
        "tags": [
          "places"
        ],
        "summary": "Suggestions",
        "description": "FR-22 — 주변 미포함 추천. 이미 담긴 장소는 제외한다.",
        "operationId": "suggestions_api_places_suggestions_get",
        "parameters": [
          {
            "name": "trip_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 36,
              "title": "Trip Id"
            }
          },
          {
            "name": "day_index",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 10,
              "minimum": 1,
              "title": "Day Index"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 50,
              "default": "맛집",
              "title": "Keyword"
            }
          },
          {
            "name": "radius",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 100,
              "default": 1500,
              "title": "Radius"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestionsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trips/{trip_id}/share": {
      "post": {
        "tags": [
          "share"
        ],
        "summary": "Issue Share Token",
        "description": "BR-36 — 암호학적 난수 43자. `trip_id` 와 수학적 관계가 없다.",
        "operationId": "issue_share_token_api_trips__trip_id__share_post",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareTokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "share"
        ],
        "summary": "Revoke Share Token",
        "description": "BR-38 — 폐기 시 기존 링크가 즉시 무효화된다.",
        "operationId": "revoke_share_token_api_trips__trip_id__share_delete",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/shared/{share_token}": {
      "get": {
        "tags": [
          "share"
        ],
        "summary": "Get Shared Trip",
        "description": "BR-37 — 편집 연산이 정의되지 않은 타입을 반환한다.\n\n응답에 `share_token` 을 포함하지 않는다 — 열람자가 토큰을 재발급·폐기할\n권한이 없기 때문이다.",
        "operationId": "get_shared_trip_api_shared__share_token__get",
        "parameters": [
          {
            "name": "share_token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 20,
              "maxLength": 64,
              "title": "Share Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadOnlyTripOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BlogRefOut": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "link": {
            "type": "string",
            "title": "Link"
          },
          "blogger_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blogger Name"
          },
          "post_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Post Date"
          }
        },
        "type": "object",
        "required": [
          "title",
          "link"
        ],
        "title": "BlogRefOut"
      },
      "BudgetLevel": {
        "type": "string",
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH"
        ],
        "title": "BudgetLevel"
      },
      "CoordinateOut": {
        "properties": {
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lng"
        ],
        "title": "CoordinateOut"
      },
...
      "DayRuleOut": {
        "properties": {
          "weekday": {
            "type": "integer",
            "title": "Weekday"
          },
          "open": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open"
          },
          "close": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close"
          },
          "closed": {
            "type": "boolean",
            "title": "Closed",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "weekday"
        ],
        "title": "DayRuleOut"
      },
      "GenerationStep": {
        "type": "string",
        "enum": [
          "DRAFTING",
          "RESOLVING",
          "ROUTING",
          "OPTIMIZING",
          "SCHEDULING",
          "SAVING"
        ],
        "title": "GenerationStep"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImageRefOut": {
        "properties": {
          "thumbnail_url": {
            "type": "string",
            "title": "Thumbnail Url"
          },
          "link": {
            "type": "string",
            "title": "Link"
          },
          "source_title": {
            "type": "string",
            "title": "Source Title",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "thumbnail_url",
          "link"
        ],
        "title": "ImageRefOut"
      },
...
      "JobAccepted": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "trip_id": {
            "type": "string",
            "title": "Trip Id"
          },
          "state": {
            "type": "string",
            "const": "queued",
            "title": "State"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "trip_id",
          "state"
        ],
        "title": "JobAccepted"
      },
      "JobState": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "succeeded",
          "partial",
          "failed"
        ],
        "title": "JobState"
      },
      "JobStatusOut": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "trip_id": {
            "type": "string",
            "title": "Trip Id"
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "step": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GenerationStep"
              },
              {
                "type": "null"
              }
            ]
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "resolved_count": {
            "type": "integer",
            "title": "Resolved Count"
          },
          "unresolved_count": {
            "type": "integer",
            "title": "Unresolved Count"
          },
          "problem": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "trip_id",
          "state",
          "progress",
          "resolved_count",
          "unresolved_count"
        ],
        "title": "JobStatusOut"
      },
...
      "OpeningHoursOut": {
        "properties": {
          "weekday_rules": {
            "items": {
              "$ref": "#/components/schemas/DayRuleOut"
            },
            "type": "array",
            "title": "Weekday Rules",
            "default": []
          },
          "entered_by_user": {
            "type": "boolean",
            "title": "Entered By User",
            "default": true
          }
        },
        "type": "object",
        "title": "OpeningHoursOut",
        "description": "BR-35 — 사용자가 입력한 경우에만 존재한다(대개 `null`)."
      },
...
      "PagedPlacesOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlaceOut"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "has_more"
        ],
        "title": "PagedPlacesOut"
      },
      "PlaceCategory": {
        "type": "string",
        "enum": [
          "RESTAURANT",
          "CAFE",
          "ATTRACTION",
          "MUSEUM",
          "SHOPPING",
          "ACCOMMODATION",
          "OTHER"
        ],
        "title": "PlaceCategory"
      },
      "PlaceContentOut": {
        "properties": {
          "place_id": {
            "type": "string",
            "title": "Place Id"
          },
          "highlights": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Highlights",
            "default": []
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/BlogRefOut"
            },
            "type": "array",
            "title": "Sources",
            "default": []
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/ImageRefOut"
            },
            "type": "array",
            "title": "Images",
            "default": []
          },
          "is_ai_summary": {
            "type": "boolean",
            "title": "Is Ai Summary",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "place_id"
        ],
        "title": "PlaceContentOut",
        "description": "FR-20·21 / BR-40 — `sources` 가 3건 미만이면 `highlights` 는 반드시 빈 목록이다."
      },
      "PlaceOut": {
        "properties": {
          "place_id": {
            "type": "string",
            "title": "Place Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "coordinate": {
            "$ref": "#/components/schemas/CoordinateOut"
          },
          "category": {
            "$ref": "#/components/schemas/PlaceCategory"
          },
          "road_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Road Address"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "category_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Raw"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "naver_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Naver Link"
          },
          "source": {
            "$ref": "#/components/schemas/PlaceSource"
          },
          "resolved_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved From"
          },
          "match_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Score"
          },
          "opening_hours": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OpeningHoursOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "place_id",
          "name",
          "coordinate",
          "category",
          "source"
        ],
        "title": "PlaceOut"
      },
      "PlaceSource": {
        "type": "string",
        "enum": [
          "NAVER_LOCAL",
          "USER_MANUAL",
          "MOCK"
        ],
        "title": "PlaceSource"
      },
...
      "ReadOnlyTripOut": {
        "properties": {
          "trip_id": {
            "type": "string",
            "title": "Trip Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "destination": {
            "type": "string",
            "title": "Destination"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "party_size": {
            "type": "integer",
            "title": "Party Size"
          },
          "style_tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Style Tags",
            "default": []
          },
          "day_start_time": {
            "type": "string",
            "format": "time",
            "title": "Day Start Time"
          },
          "day_end_time": {
            "type": "string",
            "format": "time",
            "title": "Day End Time"
          },
          "default_travel_mode": {
            "$ref": "#/components/schemas/TravelMode"
          },
          "budget_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BudgetLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "days": {
            "items": {
              "$ref": "#/components/schemas/TripDayOut"
            },
            "type": "array",
            "title": "Days"
          },
          "read_only": {
            "type": "boolean",
            "const": true,
            "title": "Read Only",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "trip_id",
          "title",
          "destination",
          "start_date",
          "end_date",
          "party_size",
          "day_start_time",
          "day_end_time",
          "default_travel_mode",
          "days"
        ],
        "title": "ReadOnlyTripOut",
        "description": "BR-37 / DD-25 — 공유 조회 응답.\n\n🔴 `share_token` 을 포함하지 않는다. 열람자는 토큰을 재발급·폐기할 권한이 없다."
      },
...
      "ShareTokenOut": {
        "properties": {
          "share_token": {
            "type": "string",
            "title": "Share Token"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "share_token",
          "url"
        ],
        "title": "ShareTokenOut"
      },
      "SuggestionsOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlaceOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "SuggestionsOut"
      },
      "TravelMode": {
        "type": "string",
        "enum": [
          "WALK",
          "CAR",
          "TRANSIT"
        ],
        "title": "TravelMode"
      },
      "TripDayOut": {
        "properties": {
          "day_index": {
            "type": "integer",
            "title": "Day Index"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItineraryItemOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "day_index",
          "items"
        ],
        "title": "TripDayOut"
      },
...
      "TripSpecIn": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Title"
          },
          "destination": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Destination"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "party_size": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Party Size",
            "default": 2
          },
          "style_tags": {
            "items": {
              "type": "string",
              "maxLength": 20
            },
            "type": "array",
            "maxItems": 8,
            "title": "Style Tags",
            "default": []
          },
          "day_start_time": {
            "type": "string",
            "format": "time",
            "title": "Day Start Time",
            "default": "09:00:00"
          },
          "day_end_time": {
            "type": "string",
            "format": "time",
            "title": "Day End Time",
            "default": "21:00:00"
          },
          "default_travel_mode": {
            "$ref": "#/components/schemas/TravelMode",
            "default": "TRANSIT"
          },
          "budget_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BudgetLevel"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "destination",
          "start_date",
          "end_date"
        ],
        "title": "TripSpecIn",
        "description": "FR-1 — 여행 생성 입력."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}